+++ test/unit/test_environment.py
@@ -57,7 +57,7 @@ SAGEMAKER_HYPERPARAMETERS = {
     'sagemaker_enable_cloudwatch_metrics': True,
     'sagemaker_container_log_level':       logging.WARNING,
     '_tuning_objective_metric':            'loss:3.4',
-    'sagemaker_channels':                  {'train': '/opt/input/data/train', 'eval': '/opt/input/data/eva/', },
+    'sagemaker_parameter_server_num':      2
 }
 
 ALL_HYPERPARAMETERS = dict(itertools.chain(USER_HYPERPARAMETERS.items(), SAGEMAKER_HYPERPARAMETERS.items()))
@@ -197,6 +197,7 @@ def test_training_env(training_env):
     assert training_env.log_level == logging.WARNING
     assert training_env.network_interface_name == 'ethwe'
     assert training_env.job_name == 'training-job-42'
+    assert training_env.additional_framework_parameters == {'sagemaker_parameter_server_num': 2}
 
 
 def test_serving_env(serving_env):
@@ -211,10 +212,10 @@ def test_serving_env(serving_env):
 
 def test_env_mapping_properties(training_env):
     assert sorted(training_env.properties()) == sorted(
-        ['channel_input_dirs', 'current_host', 'framework_module', 'hosts', 'hyperparameters', 'input_config_dir',
-         'input_data_config', 'input_dir', 'log_level', 'model_dir', 'module_dir', 'module_name',
-         'network_interface_name', 'num_cpus', 'num_gpus', 'output_data_dir', 'output_dir', 'resource_config',
-         'job_name'])
+        ['additional_framework_parameters', 'channel_input_dirs', 'current_host', 'framework_module', 'hosts',
+         'hyperparameters', 'input_config_dir', 'input_data_config', 'input_dir', 'log_level', 'model_dir',
+         'module_dir', 'module_name', 'network_interface_name', 'num_cpus', 'num_gpus', 'output_data_dir',
+         'output_dir', 'resource_config', 'job_name'])
 
 
 def test_serving_env_properties(serving_env):