+++ test/unit/test_modules.py
@@ -154,13 +154,22 @@ def test_run():
     _modules.run('pytest', ['--version'])
 
 
-def test_run_module():
+def test_run_module_wait():
     with patch('sagemaker_containers._modules.download_and_install') as download_and_install:
         with patch('sagemaker_containers._modules.run') as run:
             _modules.run_module(uri='s3://url', args=['42'], cache=True)
 
             download_and_install.assert_called_with('s3://url', 'default_user_module_name', True)
-            run.assert_called_with('default_user_module_name', ['42'], {})
+            run.assert_called_with('default_user_module_name', ['42'], {}, True)
+
+
+def test_run_module_no_wait():
+    with patch('sagemaker_containers._modules.download_and_install') as download_and_install:
+        with patch('sagemaker_containers._modules.run') as run:
+            _modules.run_module(uri='s3://url', args=['42'], cache=True, wait=False)
+
+            download_and_install.assert_called_with('s3://url', 'default_user_module_name', True)
+            run.assert_called_with('default_user_module_name', ['42'], {}, False)
 
 
 def test_download_and_install_local_directory():